fixed multiple missing processing steps for point classifications and velocities
I discovered this when I tried to import a large laszip file containing classifications, which resulted in a crash.
Further investigation showed that classification data randomly disappeared in the processing pipeline because a number of functions simply ignored classification data.
This resulted in random crashes when trying to merge different octrees with and without classifications.
The same problem existed for velocity data.